requirements.txt
streamlit
PyMuPDF
google-generativeai
streamlit-extras